chore: add more mobile apis (#5665)
