+++ packages/mask/background/database/avatar-cache/avatar.ts
@@ -1,22 +1,31 @@
 import { ProfileIdentifier } from '@masknet/shared-base'
 import { queryAvatarDB, isAvatarOutdatedDB, storeAvatarDB, IdentifierWithAvatar, createAvatarDBAccess } from './db'
 import { MaskMessages } from '../../../shared'
+import { hasNativeAPI, nativeAPI } from '../../../shared/native-rpc'
 import { blobToDataURL, memoizePromise } from '@dimensiondev/kit'
 import { createTransaction } from '../utils/openDB'
 
 /**
- * Get a (cached) blob url for an identifier.
+ * Get a (cached) blob url for an identifier. No cache for native api.
  * ? Because of cross-origin restrictions, we cannot use blob url here. sad :(
  */
-export const queryAvatarDataURL = memoizePromise(
-    async function (identifier: IdentifierWithAvatar): Promise<string | undefined> {
-        const t = createTransaction(await createAvatarDBAccess(), 'readonly')('avatars')
-        const buffer = await queryAvatarDB(t, identifier)
-        if (!buffer) throw new Error('Avatar not found')
-        return blobToDataURL(new Blob([buffer], { type: 'image/png' }))
-    },
-    (id) => id.toText(),
-)
+export const queryAvatarDataURL = (
+    hasNativeAPI
+        ? async function (identifier: IdentifierWithAvatar): Promise<string | undefined> {
+              return nativeAPI?.api.query_avatar({ identifier: identifier.toText() })
+          }
+        : memoizePromise(
+              async function (identifier: IdentifierWithAvatar): Promise<string | undefined> {
+                  const t = createTransaction(await createAvatarDBAccess(), 'readonly')('avatars')
+                  const buffer = await queryAvatarDB(t, identifier)
+                  if (!buffer) throw new Error('Avatar not found')
+                  return blobToDataURL(new Blob([buffer], { type: 'image/png' }))
+              },
+              (id) => id.toText(),
+          )
+) as ((identifier: IdentifierWithAvatar) => Promise<string | undefined>) & {
+    cache?: Map<string, unknown>
+}
 
 /**
  * Store an avatar with a url for an identifier.
@@ -28,6 +37,12 @@ export const queryAvatarDataURL = memoizePromise(
 export async function storeAvatar(identifier: IdentifierWithAvatar, avatar: ArrayBuffer | string): Promise<void> {
     if (identifier instanceof ProfileIdentifier && identifier.isUnknown) return
     try {
+        if (hasNativeAPI) {
+            // ArrayBuffer is unreachable on Native side.
+            if (typeof avatar !== 'string') return
+            await nativeAPI?.api.store_avatar({ identifier: identifier.toText(), avatar: avatar as string })
+            return
+        }
         if (typeof avatar === 'string') {
             if (avatar.startsWith('https') === false) return
             const isOutdated = await isAvatarOutdatedDB(
@@ -51,7 +66,7 @@ export async function storeAvatar(identifier: IdentifierWithAvatar, avatar: Arra
     } catch (error) {
         console.error('[AvatarDB] Store avatar failed', error)
     } finally {
-        queryAvatarDataURL.cache.delete(identifier.toText())
+        queryAvatarDataURL.cache?.delete(identifier.toText())
         if (identifier instanceof ProfileIdentifier) {
             MaskMessages.events.profilesChanged.sendToAll([{ of: identifier, reason: 'update' }])
         }
