+++ packages/mask/background/database/post/type.ts
@@ -1,0 +1,76 @@
+import type {
+    AESJsonWebKey,
+    GroupIdentifier,
+    IdentifierMap,
+    PersonaIdentifier,
+    PostIVIdentifier,
+    ProfileIdentifier,
+} from '@masknet/shared-base'
+import type { DBSchema } from 'idb/with-async-ittr'
+import type { PrototypeLess } from '../../../utils-pure'
+import type { IDBPSafeTransaction } from '../utils/openDB'
+
+export type RecipientReason = (
+    | { type: 'auto-share' }
+    | { type: 'direct' }
+    | { type: 'group'; group: GroupIdentifier }
+) & {
+    /**
+     * When we send the key to them by this reason?
+     * If the unix timestamp of this Date is 0,
+     * should display it as "unknown" or "before Nov 2019"
+     */
+    at: Date
+}
+export interface RecipientDetail {
+    /** Why they're able to receive this message? */
+    reason: RecipientReason[]
+}
+export interface PostRecord {
+    /**
+     * For old data stored before version 3, this identifier may be ProfileIdentifier.unknown
+     */
+    postBy: ProfileIdentifier
+    identifier: PostIVIdentifier
+    postCryptoKey?: AESJsonWebKey
+    /**
+     * Receivers
+     */
+    recipients: 'everyone' | IdentifierMap<ProfileIdentifier, RecipientDetail>
+    /** @deprecated */
+    recipientGroups?: unknown
+    /**
+     * When does Mask find this post.
+     * For your own post, it is when Mask created this post.
+     * For others post, it is when you see it first time.
+     */
+    foundAt: Date
+    encryptBy?: PersonaIdentifier
+    /** The URL of this post */
+    url?: string
+    /** Summary of this post (maybe front 20 chars). */
+    summary?: string
+    /** Interested metadata contained in this post. */
+    interestedMeta?: ReadonlyMap<string, unknown>
+}
+
+export interface PostDBRecord extends Omit<PostRecord, 'postBy' | 'identifier' | 'recipients' | 'encryptBy'> {
+    postBy: PrototypeLess<ProfileIdentifier>
+    identifier: string
+    recipients: true | Map<string, RecipientDetail>
+    encryptBy?: string
+}
+
+export interface PostDB extends DBSchema {
+    /** Use inline keys */
+    post: {
+        value: PostDBRecord
+        key: string
+        indexes: {
+            'persona, date': [string, Date]
+        }
+    }
+}
+
+export type PostReadOnlyTransaction = IDBPSafeTransaction<PostDB, ['post'], 'readonly'>
+export type PostReadWriteTransaction = IDBPSafeTransaction<PostDB, ['post'], 'readwrite'>
