+++ packages/public-api/src/mobile.ts
@@ -6,6 +6,7 @@ import type {
     ProfileRecord,
     LinkedProfileDetails,
     RelationRecord,
+    PostRecord,
 } from './types'
 
 /**
@@ -100,6 +101,17 @@ export interface SharedNativeAPIs {
     }): Promise<RelationRecord[]>
     update_relation(params: { relation: Omit<RelationRecord, 'network'> }): Promise<RelationRecord>
     delete_relation(params: { personaIdentifier: string; profileIdentifier: string }): Promise<void>
+    query_avatar(params: { identifier: string }): Promise<string>
+    store_avatar(params: { identifier: string; avatar: string }): Promise<void>
+    create_post(params: { post: PostRecord }): Promise<PostRecord>
+    query_post(params: { identifier: string }): Promise<PostRecord>
+    query_posts(params: {
+        encryptBy?: string
+        userIds: string[]
+        network?: string
+        pageOption?: PageOption
+    }): Promise<PostRecord[]>
+    update_post(params: { post: Partial<PostRecord>; options: { mode: 0 | 1 } }): Promise<PostRecord[]>
 }
 /**
  * APIs that only implemented by iOS Mask Network
